+++ packages/signature-controller/src/SignatureController.test.ts
@@ -1,5 +1,5 @@
 import type { SIWEMessage } from '@metamask/controller-utils';
-import { detectSIWE } from '@metamask/controller-utils';
+import { detectSIWE, ORIGIN_METAMASK } from '@metamask/controller-utils';
 import { SignTypedDataVersion } from '@metamask/keyring-controller';
 import { LogType, SigningStage } from '@metamask/logging-controller';
 import { v1 } from 'uuid';
@@ -11,7 +11,12 @@ import type {
   SignatureControllerState,
 } from './SignatureController';
 import { SignatureController } from './SignatureController';
-import type { MessageParamsPersonal, SignatureRequest } from './types';
+import type {
+  MessageParamsPersonal,
+  MessageParamsTyped,
+  OriginalRequest,
+  SignatureRequest,
+} from './types';
 import { SignatureRequestStatus, SignatureRequestType } from './types';
 import {
   normalizePersonalMessageParams,
@@ -29,20 +34,31 @@ jest.mock('@metamask/controller-utils', () => ({
 
 const ID_MOCK = '123-456';
 const CHAIN_ID_MOCK = '0x1';
+const NETWORK_CLIENT_ID_MOCK = 'testNetworkClientId';
 const FROM_MOCK = '0x456DEF';
 const DATA_MOCK = '0xABC123';
 const SIGNATURE_HASH_MOCK = '0x123ABC';
 const ERROR_MESSAGE_MOCK = 'Test Error Message';
 const ERROR_CODE_MOCK = 1234;
+const ORIGIN_MOCK = 'testOrigin';
 
 const PARAMS_MOCK = {
-  from: FROM_MOCK,
   data: DATA_MOCK,
+  from: FROM_MOCK,
+  metamaskId: ID_MOCK,
+  origin: ORIGIN_MOCK,
+  version: SignTypedDataVersion.V1,
+};
+
+const REQUEST_MOCK = {
+  networkClientId: NETWORK_CLIENT_ID_MOCK,
 };
 
 const SIGNATURE_REQUEST_MOCK: SignatureRequest = {
+  chainId: CHAIN_ID_MOCK,
   id: ID_MOCK,
   messageParams: PARAMS_MOCK,
+  networkClientId: NETWORK_CLIENT_ID_MOCK,
   status: SignatureRequestStatus.Signed,
   time: Date.now(),
   type: SignatureRequestType.PersonalSign,
@@ -57,6 +73,7 @@ function createMessengerMock() {
   const approvalControllerAddRequestMock = jest.fn();
   const keyringControllerSignPersonalMessageMock = jest.fn();
   const keyringControllerSignTypedMessageMock = jest.fn();
+  const networkControllerGetNetworkClientByIdMock = jest.fn();
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const callMock = (method: string, ...args: any[]) => {
@@ -69,6 +86,8 @@ function createMessengerMock() {
         return keyringControllerSignPersonalMessageMock(...args);
       case 'KeyringController:signTypedMessage':
         return keyringControllerSignTypedMessageMock(...args);
+      case 'NetworkController:getNetworkClientById':
+        return networkControllerGetNetworkClientByIdMock(...args);
       default:
         throw new Error(`Messenger method not recognised: ${method}`);
     }
@@ -83,6 +102,12 @@ function createMessengerMock() {
 
   approvalControllerAddRequestMock.mockResolvedValue({});
   loggingControllerAddMock.mockResolvedValue({});
+
+  networkControllerGetNetworkClientByIdMock.mockReturnValue({
+    configuration: {
+      chainId: CHAIN_ID_MOCK,
+    },
+  });
 
   return {
     approvalControllerAddRequestMock,
@@ -103,7 +128,6 @@ function createController(options?: Partial<SignatureControllerOptions>) {
 
   const controller = new SignatureController({
     messenger: messengerMocks.messenger,
-    getCurrentChainId: () => CHAIN_ID_MOCK,
     ...options,
   });
 
@@ -327,15 +351,26 @@ describe('SignatureController', () => {
   describe.each([
     [
       'newUnsignedPersonalMessage',
-      (controller: SignatureController, request = {}) =>
-        controller.newUnsignedPersonalMessage({ ...PARAMS_MOCK }, request),
+      (
+        controller: SignatureController,
+        request: OriginalRequest,
+        params?: Partial<MessageParamsPersonal>,
+      ) =>
+        controller.newUnsignedPersonalMessage(
+          { ...PARAMS_MOCK, ...params },
+          request,
+        ),
       SignatureRequestType.PersonalSign,
     ],
     [
       'newUnsignedTypedMessage',
-      (controller: SignatureController, request = {}) =>
+      (
+        controller: SignatureController,
+        request: OriginalRequest,
+        params?: Partial<MessageParamsTyped>,
+      ) =>
         controller.newUnsignedTypedMessage(
-          PARAMS_MOCK,
+          { ...PARAMS_MOCK, ...params },
           request,
           SignTypedDataVersion.V1,
           { parseJsonData: false },
@@ -352,7 +387,7 @@ describe('SignatureController', () => {
       approvalControllerAddRequestMock.mockRejectedValueOnce(error);
 
       await expect(
-        controller.newUnsignedPersonalMessage({ ...PARAMS_MOCK }, {}),
+        controller.newUnsignedPersonalMessage({ ...PARAMS_MOCK }, REQUEST_MOCK),
       ).rejects.toMatchObject({
         message: ERROR_MESSAGE_MOCK,
         code: ERROR_CODE_MOCK,
@@ -383,7 +418,7 @@ describe('SignatureController', () => {
         SIGNATURE_HASH_MOCK,
       );
 
-      await fn(controller);
+      await fn(controller, REQUEST_MOCK);
 
       expect(resultCallbackSuccessMock).toHaveBeenCalledTimes(1);
       expect(resultCallbackSuccessMock).toHaveBeenCalledWith(
@@ -398,7 +433,7 @@ describe('SignatureController', () => {
 
       controller.hub.on(`${ID_MOCK}:finished`, listener);
 
-      await fn(controller);
+      await fn(controller, REQUEST_MOCK);
 
       const state = controller.state.signatureRequests[ID_MOCK];
 
@@ -418,7 +453,7 @@ describe('SignatureController', () => {
 
       approvalControllerAddRequestMock.mockRejectedValueOnce(errorMock);
 
-      await fn(controller).catch(() => {
+      await fn(controller, REQUEST_MOCK).catch(() => {
         // Ignore error
       });
 
@@ -431,7 +466,7 @@ describe('SignatureController', () => {
     it('adds logs to logging controller if approved', async () => {
       const { controller, loggingControllerAddMock } = createController();
 
-      await fn(controller);
+      await fn(controller, REQUEST_MOCK);
 
       expect(loggingControllerAddMock).toHaveBeenCalledTimes(2);
 
@@ -465,7 +500,7 @@ describe('SignatureController', () => {
 
       approvalControllerAddRequestMock.mockRejectedValueOnce(errorMock);
 
-      await expect(fn(controller)).rejects.toThrow(errorMock);
+      await expect(fn(controller, REQUEST_MOCK)).rejects.toThrow(errorMock);
 
       expect(loggingControllerAddMock).toHaveBeenCalledTimes(2);
 
@@ -488,24 +523,52 @@ describe('SignatureController', () => {
       });
     });
 
-    it('populates origin from request', async () => {
+    it('populates origin from request if present', async () => {
       const { controller } = createController();
 
-      await fn(controller, { origin: 'test' });
+      await fn(controller, { ...REQUEST_MOCK, origin: 'test' });
 
       expect(
         controller.state.signatureRequests[ID_MOCK].messageParams.origin,
       ).toBe('test');
     });
 
+    it('populates origin from message params if no request', async () => {
+      const { controller } = createController();
+
+      await fn(controller, REQUEST_MOCK);
+
+      expect(
+        controller.state.signatureRequests[ID_MOCK].messageParams.origin,
+      ).toBe(ORIGIN_MOCK);
+    });
+
     it('populates request ID from request', async () => {
       const { controller } = createController();
 
-      await fn(controller, { id: 'test' });
+      await fn(controller, { ...REQUEST_MOCK, id: 123 });
 
       expect(
         controller.state.signatureRequests[ID_MOCK].messageParams.requestId,
-      ).toBe('test');
+      ).toBe(123);
+    });
+
+    it('populates metamask ID using ID', async () => {
+      const { controller } = createController();
+
+      await fn(controller, REQUEST_MOCK);
+
+      expect(
+        controller.state.signatureRequests[ID_MOCK].messageParams.metamaskId,
+      ).toBe(ID_MOCK);
+    });
+
+    it('throws if no network client ID in request', async () => {
+      const { controller } = createController();
+
+      await expect(fn(controller, {} as OriginalRequest)).rejects.toThrow(
+        'Network client ID not found in request',
+      );
     });
 
     it('emits unapproved message event', async () => {
@@ -515,7 +578,7 @@ describe('SignatureController', () => {
 
       controller.hub.on('unapprovedMessage', listener);
 
-      await fn(controller);
+      await fn(controller, REQUEST_MOCK);
 
       expect(listener).toHaveBeenCalledTimes(1);
       expect(listener).toHaveBeenCalledWith({
@@ -543,7 +606,7 @@ describe('SignatureController', () => {
 
       controller.hub.on(`${type}:signed`, listener);
 
-      await fn(controller);
+      await fn(controller, REQUEST_MOCK);
 
       expect(listener).toHaveBeenCalledTimes(1);
       expect(listener).toHaveBeenCalledWith({
@@ -568,7 +631,7 @@ describe('SignatureController', () => {
       keyringControllerSignPersonalMessageMock.mockRejectedValueOnce(errorMock);
       keyringControllerSignTypedMessageMock.mockRejectedValueOnce(errorMock);
 
-      await fn(controller).catch(() => {
+      await fn(controller, REQUEST_MOCK).catch(() => {
         // Ignore error
       });
 
@@ -599,10 +662,69 @@ describe('SignatureController', () => {
       keyringControllerSignPersonalMessageMock.mockRejectedValueOnce(errorMock);
       keyringControllerSignTypedMessageMock.mockRejectedValueOnce(errorMock);
 
-      await expect(fn(controller)).rejects.toThrow(ERROR_MESSAGE_MOCK);
+      await expect(fn(controller, REQUEST_MOCK)).rejects.toThrow(
+        ERROR_MESSAGE_MOCK,
+      );
 
       expect(resultCallbackErrorMock).toHaveBeenCalledTimes(1);
       expect(resultCallbackErrorMock).toHaveBeenCalledWith(errorMock);
+    });
+
+    it('requests approval', async () => {
+      const { controller, approvalControllerAddRequestMock } =
+        createController();
+
+      await fn(controller, REQUEST_MOCK);
+
+      expect(approvalControllerAddRequestMock).toHaveBeenCalledTimes(1);
+      expect(approvalControllerAddRequestMock).toHaveBeenCalledWith(
+        {
+          expectsResult: true,
+          id: ID_MOCK,
+          origin: ORIGIN_MOCK,
+          requestData: expect.objectContaining({
+            data: PARAMS_MOCK.data,
+            from: PARAMS_MOCK.from,
+          }),
+          type,
+        },
+        true,
+      );
+    });
+
+    it('requests approval with internal origin if no origin provided', async () => {
+      const { controller, approvalControllerAddRequestMock } =
+        createController();
+
+      await fn(controller, REQUEST_MOCK, { origin: undefined });
+
+      expect(approvalControllerAddRequestMock).toHaveBeenCalledTimes(1);
+      expect(approvalControllerAddRequestMock).toHaveBeenCalledWith(
+        {
+          expectsResult: true,
+          id: ID_MOCK,
+          origin: ORIGIN_METAMASK,
+          requestData: expect.objectContaining({
+            data: PARAMS_MOCK.data,
+            from: PARAMS_MOCK.from,
+          }),
+          type,
+        },
+        true,
+      );
+    });
+
+    it('persists network client ID and chain ID in metadata', async () => {
+      const { controller } = createController();
+
+      await fn(controller, REQUEST_MOCK);
+
+      expect(controller.state.signatureRequests[ID_MOCK]).toStrictEqual(
+        expect.objectContaining({
+          chainId: CHAIN_ID_MOCK,
+          networkClientId: NETWORK_CLIENT_ID_MOCK,
+        }),
+      );
     });
   });
 
@@ -617,7 +739,7 @@ describe('SignatureController', () => {
 
       const result = await controller.newUnsignedPersonalMessage(
         { ...PARAMS_MOCK },
-        {},
+        REQUEST_MOCK,
       );
 
       expect(result).toBe(SIGNATURE_HASH_MOCK);
@@ -633,7 +755,10 @@ describe('SignatureController', () => {
 
       detectSIWEMock.mockReturnValueOnce(siweMock);
 
-      await controller.newUnsignedPersonalMessage({ ...PARAMS_MOCK }, {});
+      await controller.newUnsignedPersonalMessage(
+        { ...PARAMS_MOCK },
+        REQUEST_MOCK,
+      );
 
       const messageParams = controller.state.signatureRequests[ID_MOCK]
         .messageParams as MessageParamsPersonal;
@@ -653,7 +778,7 @@ describe('SignatureController', () => {
 
       const result = await controller.newUnsignedTypedMessage(
         PARAMS_MOCK,
-        {},
+        REQUEST_MOCK,
         SignTypedDataVersion.V1,
         { parseJsonData: false },
       );
@@ -676,7 +801,7 @@ describe('SignatureController', () => {
             ...PARAMS_MOCK,
             data: JSON.stringify({ test: 123 }),
           },
-          {},
+          REQUEST_MOCK,
           version as SignTypedDataVersion,
           { parseJsonData: true },
         );
@@ -685,6 +810,7 @@ describe('SignatureController', () => {
           {
             ...PARAMS_MOCK,
             data: { test: 123 },
+            version,
           },
           version,
         );
@@ -706,7 +832,7 @@ describe('SignatureController', () => {
           ...PARAMS_MOCK,
           data: JSON.stringify({ test: 123 }),
         },
-        {},
+        REQUEST_MOCK,
         SignTypedDataVersion.V1,
         { parseJsonData: true },
       );
@@ -733,7 +859,7 @@ describe('SignatureController', () => {
       await expect(
         controller.newUnsignedTypedMessage(
           PARAMS_MOCK,
-          {},
+          REQUEST_MOCK,
           SignTypedDataVersion.V3,
           { parseJsonData: false },
         ),
@@ -745,6 +871,24 @@ describe('SignatureController', () => {
       expect(controller.state.signatureRequests[ID_MOCK].error).toBe(
         ERROR_MESSAGE_MOCK,
       );
+    });
+
+    it('populates version in params', async () => {
+      const { controller } = createController();
+
+      await controller.newUnsignedTypedMessage(
+        PARAMS_MOCK,
+        REQUEST_MOCK,
+        SignTypedDataVersion.V3,
+        { parseJsonData: false },
+      );
+
+      expect(
+        (
+          controller.state.signatureRequests[ID_MOCK]
+            .messageParams as MessageParamsTyped
+        ).version,
+      ).toBe(SignTypedDataVersion.V3);
     });
   });
 
@@ -782,7 +926,7 @@ describe('SignatureController', () => {
             ...PARAMS_MOCK,
             deferSetAsSigned: true,
           },
-          {},
+          REQUEST_MOCK,
         )
         .then((result) => {
           resolved = true;
@@ -860,7 +1004,7 @@ describe('SignatureController', () => {
             ...PARAMS_MOCK,
             deferSetAsSigned: true,
           },
-          {},
+          REQUEST_MOCK,
         )
         .catch((error) => {
           rejectedError = error;