+++ packages/signature-controller/src/types.ts
@@ -1,11 +1,14 @@
 import type { SIWEMessage } from '@metamask/controller-utils';
 import type { SignTypedDataVersion } from '@metamask/keyring-controller';
-import type { Json } from '@metamask/utils';
+import type { Hex, Json } from '@metamask/utils';
 
 /** Original client request that triggered the signature request. */
 export type OriginalRequest = {
   /** Unique ID to identify the client request. */
   id?: number;
+
+  /** ID of the network client associated with the request. */
+  networkClientId?: string;
 
   /** Source of the client request. */
   origin?: string;
@@ -27,6 +30,9 @@ export type MessageParams = {
 
   /** Ethereum address to sign with. */
   from: string;
+
+  /** ID of the associated signature request. */
+  metamaskId?: string;
 
   /**
    * Source of the request.
@@ -65,9 +71,15 @@ export type MessageParamsTyped = MessageParams & {
         primaryType: string;
         message: Json;
       };
+
+  /** Version of the signTypedData request. */
+  version?: string;
 };
 
 type SignatureRequestBase = {
+  /** ID of the associated chain. */
+  chainId: Hex;
+
   /** Error message that occurred during the signing. */
   error?: string;
 
@@ -76,6 +88,9 @@ type SignatureRequestBase = {
 
   /** Custom metadata stored with the request. */
   metadata?: Json;
+
+  /** ID of the associated network client. */
+  networkClientId: string;
 
   /** Signature hash resulting from the request. */
   rawSig?: string;